Allow 'State Admin Read Only' to view Camera
